Fix compilation issue in  ASAN machines

+++ src/btree/bt_page.c
@@ -126,7 +126,7 @@ err:
 static bool
 __is_delta_cell_type_visible_all(WT_CELL_UNPACK_DELTA_INT *unpack_delta)
 {
-    uint8_t cell_type;
+    u_int cell_type;
 
     cell_type = __wt_cell_type_raw(unpack_delta->value.cell);
     return (cell_type == WT_CELL_ADDR_DEL_VISIBLE_ALL);
